# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bisiklet Takip Uygulaması API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Bike {
  id: number;
  name: string;
  brand: string;
  model: string;
  /** road, mountain, hybrid, electric, gravel */
  type: string;
  /** @nullable */
  color: string | null;
  /** @nullable */
  purchaseDate: string | null;
  totalKm: number;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBikeBody {
  name: string;
  brand: string;
  model: string;
  type: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateBikeBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Ride {
  id: number;
  bikeId: number;
  /** @nullable */
  bikeName: string | null;
  title: string;
  date: string;
  distanceKm: number;
  durationMinutes: number;
  /** @nullable */
  avgSpeedKmh: number | null;
  /** @nullable */
  maxSpeedKmh: number | null;
  /** @nullable */
  elevationGainM: number | null;
  /** @nullable */
  calories: number | null;
  /** @nullable */
  notes: string | null;
  /**
   * city, mountain, trail, road
   * @nullable
   */
  routeType: string | null;
  createdAt: string;
}

export interface CreateRideBody {
  bikeId: number;
  title: string;
  date: string;
  distanceKm: number;
  durationMinutes: number;
  /** @nullable */
  avgSpeedKmh?: number | null;
  /** @nullable */
  maxSpeedKmh?: number | null;
  /** @nullable */
  elevationGainM?: number | null;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  routeType?: string | null;
}

export interface UpdateRideBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  distanceKm?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  avgSpeedKmh?: number | null;
  /** @nullable */
  maxSpeedKmh?: number | null;
  /** @nullable */
  elevationGainM?: number | null;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  routeType?: string | null;
}

export interface MaintenanceRecord {
  id: number;
  bikeId: number;
  /** @nullable */
  bikeName: string | null;
  /** tire, chain, brake, gear, service, other */
  type: string;
  description: string;
  date: string;
  /** @nullable */
  cost: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CreateMaintenanceBody {
  bikeId: number;
  type: string;
  description: string;
  date: string;
  /** @nullable */
  cost?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface StatsSummary {
  totalRides: number;
  totalKm: number;
  totalMinutes: number;
  totalCalories: number;
  avgSpeedKmh: number;
  longestRideKm: number;
  thisMonthKm: number;
  thisMonthRides: number;
  totalBikes: number;
}

export interface WeeklyStat {
  week: string;
  rides: number;
  km: number;
  minutes: number;
}

export interface MonthlyStat {
  month: string;
  rides: number;
  km: number;
  minutes: number;
}

export type ListRidesParams = {
  /**
   * @nullable
   */
  bikeId?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};

export type ListMaintenanceParams = {
  /**
   * @nullable
   */
  bikeId?: number | null;
};