# lib/api-zod/src/generated/types/bike.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bisiklet Takip Uygulaması API
 * OpenAPI spec version: 0.1.0
 */

export interface Bike {
  id: number;
  name: string;
  brand: string;
  model: string;
  /** road, mountain, hybrid, electric, gravel */
  type: string;
  /** @nullable */
  color: string | null;
  /** @nullable */
  purchaseDate: string | null;
  totalKm: number;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}