# lib/api-zod/src/generated/types/maintenanceRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bisiklet Takip Uygulaması API
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenanceRecord {
  id: number;
  bikeId: number;
  /** @nullable */
  bikeName: string | null;
  /** tire, chain, brake, gear, service, other */
  type: string;
  description: string;
  date: string;
  /** @nullable */
  cost: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}