# lib/api-zod/src/generated/types/ride.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bisiklet Takip Uygulaması API
 * OpenAPI spec version: 0.1.0
 */

export interface Ride {
  id: number;
  bikeId: number;
  /** @nullable */
  bikeName: string | null;
  title: string;
  date: string;
  distanceKm: number;
  durationMinutes: number;
  /** @nullable */
  avgSpeedKmh: number | null;
  /** @nullable */
  maxSpeedKmh: number | null;
  /** @nullable */
  elevationGainM: number | null;
  /** @nullable */
  calories: number | null;
  /** @nullable */
  notes: string | null;
  /**
   * city, mountain, trail, road
   * @nullable
   */
  routeType: string | null;
  createdAt: string;
}