# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bisiklet Takip Uygulaması API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalRides: number;
  totalKm: number;
  totalMinutes: number;
  totalCalories: number;
  avgSpeedKmh: number;
  longestRideKm: number;
  thisMonthKm: number;
  thisMonthRides: number;
  totalBikes: number;
}