# lib/api-zod/src/generated/types/updateBikeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bisiklet Takip Uygulaması API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBikeBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  notes?: string | null;
}