# lib/api-zod/src/generated/types/updateRideBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bisiklet Takip Uygulaması API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateRideBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  distanceKm?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  avgSpeedKmh?: number | null;
  /** @nullable */
  maxSpeedKmh?: number | null;
  /** @nullable */
  elevationGainM?: number | null;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  routeType?: string | null;
}