# lib/api-zod/src/generated/types/weeklyStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bisiklet Takip Uygulaması API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyStat {
  week: string;
  rides: number;
  km: number;
  minutes: number;
}